refactor: LogAspect 내부 클래스 외부로 추출, static 메소드 삭제
